Added unit test fixture for /simple_get_with_body
