+++ lib/http_server_test_fixture/mock_handlers/get_with_body.ex
@@ -1,0 +1,9 @@
+defmodule HTTPServerTestFixture.Handlers.GetWithBody do
+  alias HTTPServer.Request
+  @behaviour HTTPServer.Handler
+
+  @impl HTTPServer.Handler
+  def handle(%Request{method: "GET"} = _req) do
+    {200, "this is a mocked get with a body"}
+  end
+end
